+++ LibrarySort/Source.cpp
@@ -313,7 +313,7 @@ void Source::sort(int column)
 
 void Source::shellsort(std::string * toSort, size_t low, size_t high)
 {
-	long long i, j, k;
+	size_t i, j, k;
 
 	sorted_col_string = toSort;
 	sorted_col_type = string;
@@ -332,7 +332,7 @@ void Source::shellsort(std::string * toSort, size_t low, size_t high)
 
 void Source::shellsort(int * toSort, size_t low, size_t high)
 {
-	long long i, j, k;
+	size_t i, j, k;
 
 	sorted_col_int = toSort;
 	sorted_col_type = inte;
